feat: 3D universe visualization with Three.js
Phase 1: Simplified Architecture
- Removed XState journey machine and hooks
- Deleted CanvasZone, StoryScene, TransitionWrapper infrastructure
- Converted scene components to simple page components
- Replaced journey orchestration with React Router navigation
- Updated CanvasApp.tsx to use BrowserRouter with route-based navigation

Phase 2: Three.js Integration
- Installed three, @react-three/fiber, @react-three/drei, @types/three
- Created DataUniverse component with:
  - Central sphere representing total emissions (size based on kg CO₂)
  - Orbiting activity spheres with realistic orbital motion
  - Color coding based on carbon intensity (green/amber/red)
  - Hover tooltips showing emission data
  - Click handlers for activity selection
  - OrbitControls for camera interaction
  - Starfield background (5000 stars)
- Integrated DataUniverse into ExplorePage as new "3D Universe" mode
- Maintained existing Timeline and Comparison visualizations

Components Added:
- src/components/viz/DataUniverse.tsx (3D visualization)
- src/components/system/Transition.tsx (simple replacement for canvas transitions)
- src/pages/WelcomePage.tsx (simplified from OnboardingScene)
- src/pages/CalculatorPage.tsx (simplified from BaselineScene)
- src/pages/ExplorePage.tsx (with 3D universe integration)
- src/pages/InsightsPage.tsx (simplified from InsightScene)

Components Removed:
- All canvas infrastructure (CanvasZone, StoryScene, SceneNavigation)
- XState journey machine and hooks
- Old scene components (OnboardingScene, BaselineScene, ExploreScene, InsightScene)
- Example files (CanvasExample, JourneyExample)

Build Status:
- ✓ TypeScript compilation successful
- ✓ All imports resolved
- ✓ Bundle builds successfully (5.47s)

Next Steps:
- ProfileContext to Zustand migration (remaining task)
- Add camera choreography (zoom animations, fly-to effects)
- Add raycasting interactions (click detection, hover glow)
- Implement 2D overlays on 3D scenes (insights panels)
