+++ apps/carbon-acx-web/package.json
@@ -21,9 +21,10 @@
     "@radix-ui/react-slot": "^1.1.0",
     "@radix-ui/react-tabs": "^1.1.1",
     "@radix-ui/react-tooltip": "^1.1.2",
+    "@react-three/drei": "^10.7.6",
+    "@react-three/fiber": "^9.4.0",
     "@tanstack/react-query": "^5.90.5",
     "@tanstack/react-virtual": "^3.0.1",
-    "@xstate/react": "^6.0.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "echarts": "^6.0.0",
@@ -36,7 +37,7 @@
     "recharts": "^2.12.7",
     "swr": "^2.3.3",
     "tailwind-merge": "^2.5.3",
-    "xstate": "^5.23.0",
+    "three": "^0.180.0",
     "zustand": "^4.5.4"
   },
   "devDependencies": {
@@ -47,6 +48,7 @@
     "@types/react": "^18.3.11",
     "@types/react-dom": "^18.3.0",
     "@types/recharts": "^2.0.1",
+    "@types/three": "^0.180.0",
     "@vitejs/plugin-react": "^4.3.4",
     "autoprefixer": "^10.4.20",
     "postcss": "^8.4.47",
